requirements.txt
aiohttp~=3.9.3
pandas~=2.2.1
PyYAML~=6.0.1
numpy~=1.26.4
streamlit~=1.32.2
plotly~=5.20.0
toml~=0.10.2
matplotlib~=3.8.3
seaborn~=0.13.2
streamlit-aggrid==0.3.5
boto3~=1.34.76
requests~=2.31.0
botocore~=1.34.76